validate trajectory bebore writing in trajectory file handler
