+++ niryo_one_commander/src/niryo_one_commander/parameters_validation.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
-# validate_params.py
-# Copyright (C) 2017 Niryo
+# parameters_validation.py
+# Copyright (C) 2018 Niryo
 # All rights reserved.
 #
 # This program is free software: you can redistribute it and/or modify
@@ -20,62 +20,15 @@
 
 
 import rospy 
-from math import sqrt
-
+from math import sqrt 
 from niryo_one_commander.robot_commander_exception import RobotCommanderException
 from niryo_one_commander.command_status import CommandStatus
-from niryo_one_commander.command_type import CommandType
 
-class ValidateParameters(): 
+class ParametersValidation(): 
 
-    def __init__(self, validation, position_manager, trajectory_manager):
+    def __init__(self, validation):
         self.validation = validation 
-        self.pos_manager = position_manager
-        self.trajectory_manager = trajectory_manager
-    
-    def validate_params(self, cmd): 
-        cmd_type = cmd.cmd_type
-        if cmd_type == CommandType.JOINTS:
-            self.validate_joints(cmd.joints)
-        elif cmd_type == CommandType.POSE:
-            self.validate_position(cmd.position)
-            self.validate_orientation(cmd.rpy)
-        elif cmd_type == CommandType.POSITION:
-            self.validate_position(cmd.position)
-        elif cmd_type == CommandType.RPY:
-            self.validate_orientation(cmd.rpy)
-        elif cmd_type == CommandType.SHIFT_POSE:
-            self.validate_shift_pose(cmd.shift)
-        elif cmd_type == CommandType.EXECUTE_TRAJ:
-            self.validate_trajectory(cmd.Trajectory)
-        elif cmd_type == CommandType.TOOL:
-            self.validate_tool_command(cmd.tool_cmd)
-        elif cmd_type == CommandType.POSE_QUAT:
-            self.validate_position(cmd.pose_quat.position)
-            self.validate_orientation_quaternion(cmd.pose_quat.orientation)
-        elif cmd_type == CommandType.SAVED_POSITION: 
-            self.validate_saved_position(cmd.saved_position_name)
-        elif CommandType.SAVED_TRAJECTORY:
-            self.validate_saved_trajectory(cmd)
-
-        else:
-            raise RobotCommanderException(CommandStatus.INVALID_PARAMETERS, "Wrong command type")
-
-
-    def validate_saved_trajectory(self, cmd): 
-        rospy.loginfo("Checking saved trajectory validity")
-        saved_traj = self.trajectory_manager.get_trajectory(cmd.saved_trajectory_id)
-        if saved_traj == None :
-            raise RobotCommanderException(CommandStatus.INVALID_PARAMETERS, "Saved trajectory  not found") 
-        self.validate_trajectory(saved_traj.trajectory_plan)
-         
-
-    def validate_saved_position(self, position_name):
-        rospy.loginfo("Checking joints validity")
-        saved_position = self.pos_manager.get_position(position_name)
-        if saved_position == None :
-            raise RobotCommanderException(CommandStatus.INVALID_PARAMETERS, "Saved position not found") 
-        self.validate_joints(saved_position.joints)
+   
           
     def validate_trajectory(self, plan):
         rospy.loginfo("Checking trajectory validity")
